remove astnode related method to HiveASTNodeUtil
